Create the first version
I have it installed on my local laptop and dev-vm. I'm pretty happy with
it.

But it's not perfect:

When run through systemd, new terminals launch in a new window rather
than a new tab.

Completely disconnecting then reconnecting to a remote does not recreate
the remote's socket until restart. SSH offers a solution
(`StreamLocalBindUnlink`), but the solution is bugged
https://bugzilla.mindrot.org/show_bug.cgi?id=2601